Partial commit of the admin code for wells
This doesn't work as expected for some reason.  The `NodeInline` is
displaying the `well` and `order` fields and nothing else.  Not sure
what I'm missing here, but figure I'll commit it and let other people
take a look.  I've narrowed it down to being something too obvious to
even notice. :-)