Reorder requirements and add base requirements
This follows more closely with the Pinax style of using a requirements
directory with multiple files.

`base.txt` contains all of the requirements that are needed for the
project.  I want to make the `setup.py` file parse this for its
dependencies when building packages.

`dev.txt` contains all of the pieces needed if you want to do
development work such as running tests and so on.